Check for ANY commotion interfaces.
If multiple mesh interfaces exist with multiple proto's we want to make sure we show the status of any that we can. The name commotionMesh is only set for the initial mesh interface. If a user creates multiple mesh interfaces, or deletes and recreates their mesh interface it will have another name. The uci:foreach stanza here identifies adhoc wireless interfaces and checks their protocols.